Struct aurae_client::AuraeClient

pub struct AuraeClient { /* private fields */ }

Instance of a single client for an Aurae consumer.

Implementations

impl AuraeClient

pub async fn default() -> Result<Self>

pub async fn new(__arg0: AuraeConfig) -> Result<Self>

Create a new AuraeClient.

Note: A new client is required for every independent execution of this process.